+++ tests/monotouch-test/MetalPerformanceShaders/MPSImageBatchTests.cs
@@ -52,7 +52,7 @@ namespace MonoTouchFixtures.MetalPerformanceShaders {
 		public void MPSImageBatchResourceSizeTest ()
 		{
 			var size = MPSImageBatch.GetResourceSize (cache);
-			Assert.That (size, Is.GreaterThan (0), "idx");
+			Assert.That (size, Is.GreaterThan ((nuint) 0), "idx");
 		}
 
 		// TODO: Enable once'MPSImageBatch.Iterate' is enabled, rdar://47282304.
