+++ tests/monotouch-test/MetalPerformanceShaders/MPSStateResourceListTests.cs
@@ -54,7 +54,7 @@ namespace MonoTouchFixtures.MetalPerformanceShaders {
 			Assert.NotNull (resList, "resList");
 
 			var state = new MPSState (device, resList);
-			Assert.That (state.ResourceCount, Is.EqualTo (10), "ResourceCount");
+			Assert.That (state.ResourceCount, Is.EqualTo ((nuint) 10), "ResourceCount");
 		}
 
 		[Test]
@@ -64,10 +64,10 @@ namespace MonoTouchFixtures.MetalPerformanceShaders {
 			Assert.NotNull (resList, "resList");
 
 			var state = new MPSState (device, resList);
-			Assert.That (state.ResourceCount, Is.EqualTo (6), "ResourceCount");
+			Assert.That (state.ResourceCount, Is.EqualTo ((nuint) 6), "ResourceCount");
 
-			Assert.That (state.GetBufferSize (5), Is.EqualTo (241), "resList[5] = 241");
-			Assert.That (state.GetBufferSize (2), Is.EqualTo (3), "resList[2] = 3");
+			Assert.That (state.GetBufferSize (5), Is.EqualTo ((nuint) 241), "resList[5] = 241");
+			Assert.That (state.GetBufferSize (2), Is.EqualTo ((nuint) 3), "resList[2] = 3");
 		}
 	}
 }
